int instead of long int

+++ rsa.c
@@ -8,13 +8,13 @@
 #define MAX_MENSAJE     100
 
 typedef struct Clave_publica {
-    unsigned long int n;
-    unsigned long int e;
+    unsigned int n;
+    unsigned int e;
 } Clave_publica;
 
 typedef struct Clave_privada {
-    unsigned long int n;
-    unsigned long int d;
+    unsigned int n;
+    unsigned int d;
 } Clave_privada;
 
 typedef struct Claves {
@@ -23,19 +23,19 @@ typedef struct Claves {
 } Claves;
 
 Claves genera_claves();
-int es_primo(unsigned long int num);
-unsigned long int encuentra_d(unsigned long int e, unsigned long int phi);
+int es_primo(unsigned int num);
+unsigned int encuentra_d(unsigned int e, unsigned int phi);
 void salida(char * mensaje);
-unsigned long int mcd(unsigned long int a, unsigned long int b);
-long int * cifra(long int * mensaje, Clave_publica clave);
-long int * descifra(long int * mensaje, Clave_privada clave);
-long int exponente_modular(long int base, long int exp, long int mod);
+unsigned int mcd(unsigned int a, unsigned int b);
+int * cifra(int * mensaje, Clave_publica clave);
+int * descifra(int * mensaje, Clave_privada clave);
+int exponente_modular(int base, int exp, int mod);
 
 /** Funcion mcd: máximo común divisor
  *  Entradas: a y b, dos números enteros
  *  Salida: mcd(a,b).
  */
-unsigned long int mcd(unsigned long int a, unsigned long int b) {
+unsigned int mcd(unsigned int a, unsigned int b) {
     while(a!=b) {
         if(a > b)
             a-=b;
@@ -49,7 +49,7 @@ unsigned long int mcd(unsigned long int a, unsigned long int b) {
  *  Entradas: base, exponente y módulo
  *  Salida: (base elevado a exp) módulo mod
  */
-long int exponente_modular(long int base, long int exp, long int mod) {
+int exponente_modular(int base, int exp, int mod) {
     int k=0;
     int ret=1;
     for(;k<exp;k++) {
@@ -72,7 +72,7 @@ void salida(char * mensaje) {
     exit(1);
 }
 
-int es_primo(unsigned long int num) {
+int es_primo(unsigned int num) {
     double max=sqrt(num);
     if(num%2==0)
         return num==2;
@@ -83,10 +83,10 @@ int es_primo(unsigned long int num) {
     return 1;
 }
 
-unsigned long int encuentra_d(unsigned long int e, unsigned long int phi) {
+unsigned int encuentra_d(unsigned int e, unsigned int phi) {
     
     // Encuentra d tal que (d * e) % phi(n) = 1
-    unsigned long int d=1;
+    unsigned int d=1;
     while(1) {
         if(((d*e) % phi) == 1)
             return d;
@@ -97,29 +97,29 @@ unsigned long int encuentra_d(unsigned long int e, unsigned long int phi) {
 
 
 Claves genera_claves() {
-    unsigned long int p, q, n, phi;
+    unsigned int p, q, n, phi;
     Claves ret[MAX_CLAVES];
     char respuesta;
     int indice=0;
 
     printf("Necesitamos dos números primos:\n");
-    scanf("%ld", &p);
+    scanf("%i", &p);
     if(!es_primo(p))
         salida("Entrada incorrecta. P no es primo");
-    scanf("%ld", &q);
+    scanf("%i", &q);
     if(!es_primo(q))
         salida("Entrada incorrecta. Q no es primo");
     if(p==q)
         salida("Entrada incorrecta. P es igual a Q");
 
     n = p * q; 
-    printf("N = %ld, p = %ld, q = %ld\n", n, p, q);
+    printf("N = %i, p = %i, q = %i\n", n, p, q);
     phi = (p-1) * (q-1); 
-    printf("Phi(n) = %ld\n", phi);
+    printf("Phi(n) = %i\n", phi);
 
     /* Elegir e tal que 1 < e < phi(n), y tal que
      * e y n sean coprimos. */
-    long int i=2,j=0;
+    int i=2,j=0;
     for(; i<phi && j<MAX_CLAVES; i++) {
         if((mcd(i,n)==1) && (mcd(i,phi)==1)){ // e y n son coprimos, y e y phi(n) también
             ret[j].publica.e=i;
@@ -132,8 +132,8 @@ Claves genera_claves() {
 
     printf("Claves generadas.\n");
     for(i=0; i<j; i++) {
-        printf("La clave pública número %ld es (e,n) = (%ld,%ld)\n", i, ret[i].publica.e, ret[i].publica.n);
-        printf("La clave privada número %ld es (d,n) = (%ld,%ld)\n", i, ret[i].privada.d, ret[i].privada.n);
+        printf("La clave pública número %i es (e,n) = (%i,%i)\n", i, ret[i].publica.e, ret[i].publica.n);
+        printf("La clave privada número %i es (d,n) = (%i,%i)\n", i, ret[i].privada.d, ret[i].privada.n);
     }
 
     while(respuesta!='s' && respuesta!='n') {
@@ -145,7 +145,7 @@ Claves genera_claves() {
     if(respuesta=='s') {
         getchar();
         while(indice<1 || indice>i) {
-            printf("Por favor escoja un número de clave válido: 1,...,%ld:\n", i-1);
+            printf("Por favor escoja un número de clave válido: 1,...,%i:\n", i-1);
             scanf("%i", &indice);
             printf("Elegido: %i\n", indice);
         }
@@ -153,14 +153,14 @@ Claves genera_claves() {
         ret[indice] = ret[0];
         ret[0] = tmp;
     }
-    printf("\n\nClave elegida:\n\n\tkp = ( %ld, %ld )\n\n\tks = ( %ld, %ld )\n\n", ret[0].publica.e, ret[0].publica.n, ret[0].privada.d, ret[0].privada.n );
+    printf("\n\nClave elegida:\n\n\tkp = ( %i, %i )\n\n\tks = ( %i, %i )\n\n", ret[0].publica.e, ret[0].publica.n, ret[0].privada.d, ret[0].privada.n );
 
     return ret[0];
 }
 
-long int * cifra(long int * mensaje, Clave_publica clave) {
+int * cifra(int * mensaje, Clave_publica clave) {
     int k=0;
-    long int *texto_cifrado = malloc(sizeof(long int) * MAX_MENSAJE);
+    int *texto_cifrado = malloc(sizeof(int) * MAX_MENSAJE);
 
     for(; mensaje[k] != -1; k++)
        texto_cifrado[k] = exponente_modular(mensaje[k], clave.e, clave.n);
@@ -170,9 +170,9 @@ long int * cifra(long int * mensaje, Clave_publica clave) {
     return texto_cifrado;
 }
 
-long int * descifra(long int * mensaje, Clave_privada clave) {
+int * descifra(int * mensaje, Clave_privada clave) {
     int k=0;
-    long int *texto_plano = malloc(sizeof(long int) * MAX_MENSAJE);
+    int *texto_plano = malloc(sizeof(int) * MAX_MENSAJE);
     
     for(; mensaje[k] != -1; k++)
         texto_plano[k] = exponente_modular(mensaje[k], clave.d, clave.n);
@@ -185,8 +185,8 @@ long int * descifra(long int * mensaje, Clave_privada clave) {
 int main() {
    Claves mis_claves = genera_claves();
    char mensaje[MAX_MENSAJE];
-   long int *cifrado, *descifrado;
-   long int *texto = malloc(sizeof(long int) * MAX_MENSAJE);
+   int *cifrado, *descifrado;
+   int *texto = malloc(sizeof(int) * MAX_MENSAJE);
    int k=0;
    printf("Escriba su mensaje:\n");
    fflush(stdin);
@@ -200,12 +200,12 @@ int main() {
    cifrado=cifra(texto, mis_claves.publica);
    printf("\n\nEl mensaje cifrado es:\n\n\t");
    for(k=0; cifrado[k] != -1; k++)
-       printf("%c", (int) cifrado[k]);
+       printf("%c", cifrado[k]);
    printf("\n");
    descifrado= descifra(cifrado, mis_claves.privada);
    printf("\n\nEl mensaje descifrado es:\n\n\t");
    for(k=0; descifrado[k] != -1; k++)
-       printf("%c", (int) descifrado[k]);
+       printf("%c", descifrado[k]);
    printf("\n\n");
    free(texto);
    free(cifrado);
